Started React frontend
Made\ a\ frontend\ in\ react\ that\ has\ a\ top\ navigation\ bar,\ a\ button\ for\ adding\ new\ items\ to\ the\ todo-list,\ and\ a\ display\ table\ for\ the\ todos.\ So\ far,\ we\ are\ storing\ all\ data\ in\ the\ React\ Components\ themselves,\ and\ are\ not\ communicating\ with\ our\ backend\ at\ all.\ All\ frontend\ code\ can\ be\ found\ in\ the\ client\ folder.\ I\ also\ added\ a\ few\ scripts\ for\ running\ the\ server\ and\ client\ both\ separately\ and\ simultaneously.\ I\ modified\ the\ main\ projects\ package.json\ to\ include\ a\ , , and  script. The  script runs everything simultaneously using the concurrently package.