Added First Redux things
Wow. Adding redux to this was a pain in ways I lack the vocabulary to articulate. I relied incredibly hard on the Traversy Tutorial for this section; without it I would have gotten nowhere. In this commit, I npm installed redux and react-redux, moved state up to the AppWide reducers, and added actions for CRUD operations on the todo list items. As of yet, we still haven't connected to the backend, but I'll get to this tommorow. Oh, and I added a modal sheet input as well. In sum, learning JS + React + Redux + Node during spring break was more ambitious than I originally anticipated, but I can't wait to see how this project turns out / how far I can get! Ad meloria!