Connected React to Backend
Here, I redid some of the redux to accomodate axios http requests to the backend of our application. I added an  ITEMS_LOADING action type so that the application could know when it was in a loading state or not, and also used thunk to allow for dipatch() calls from within the action creator functions of our Redux stuff. This is now a working to-do app, but now is where I plan on adding more features (authentication, timed reminders possibly?) and maybe improving the UI as well